+++ modules/@angular/core/testing.ts
@@ -9,5 +9,6 @@
 export * from './testing/async';
 export * from './testing/component_fixture';
 export * from './testing/fake_async';
+export * from './testing/test_component_builder';
 export * from './testing/test_injector';
 export * from './testing/testing';
