+++ modules/@angular/platform-browser-dynamic/testing.ts
@@ -7,7 +7,8 @@
  */
 
 import {DirectiveResolver, ViewResolver} from '@angular/compiler';
-import {MockDirectiveResolver, MockViewResolver, TestComponentBuilder, TestComponentRenderer} from '@angular/compiler/testing';
+import {MockDirectiveResolver, MockViewResolver, OverridingTestComponentBuilder} from '@angular/compiler/testing';
+import {TestComponentBuilder, TestComponentRenderer} from '@angular/core/testing';
 import {TEST_BROWSER_APPLICATION_PROVIDERS, TEST_BROWSER_PLATFORM_PROVIDERS} from '@angular/platform-browser/testing';
 
 import {BROWSER_APP_COMPILER_PROVIDERS} from './index';
@@ -27,9 +28,9 @@ export const TEST_BROWSER_DYNAMIC_PLATFORM_PROVIDERS: Array<any /*Type | Provide
 export const TEST_BROWSER_DYNAMIC_APPLICATION_PROVIDERS: Array<any /*Type | Provider | any[]*/> = [
   TEST_BROWSER_APPLICATION_PROVIDERS, BROWSER_APP_COMPILER_PROVIDERS,
   [
+    {provide: TestComponentBuilder, useClass: OverridingTestComponentBuilder},
     {provide: DirectiveResolver, useClass: MockDirectiveResolver},
     {provide: ViewResolver, useClass: MockViewResolver},
-    TestComponentBuilder,
     {provide: TestComponentRenderer, useClass: DOMTestComponentRenderer},
   ]
 ];
